requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langgraph==0.0.26
snowflake-connector-python[pandas]==3.7.0
pydantic==2.6.3
python-dotenv==1.0.1
pyspark==3.5.1
